Add files via upload
To learn about the two most important types of layers that are used in the base of a convolutional image classifier. These are the convolutional layer with ReLU activation and the maximum pooling layer. In Lesson 5, you'll learn how to design your own convnet by composing these layers into blocks that perform the feature extraction.

This lesson is about the convolutional layer with its ReLU activation function.