Add files via upload
In the previous two lessons, we learned about the three operations that carry out feature extraction from an image:

filter with a convolution layer
detect with ReLU activation
condense with a maximum pooling layer
The convolution and pooling operations share a common feature: they are both performed over a sliding window. With convolution, this "window" is given by the dimensions of the kernel, the parameter kernel_size. With pooling, it is the pooling window, given by pool_size.